Completed run to position for arm
Fixed and tested various aspects of the arm running to position.  It can execute a sequence of arm movements at full power, but movements are jerky.  Not all positions are tested and much tuning is required.  To move more smoothly we have to run to velocity and control the speed or reduce the power for running to position.  Movements are too jerky for precise control unless you wait for vibrations to stop before continuing.