Clean up plus Added GamepadController2
The addition of GamePadController2 comes with an infrastructure for navigating the backdrop cells for pixel placement.  None of this has been tested yet.

There was some cleaning up of code fragments that were not in use throughout the other classes.  The addition of a boolean to track if we are in autonomous or not, so that we know to start an autonomous sequence or not in the autopilot.