Added collision avoidance and starting field coordinates for the driver op
Collision avoidance was created to avoid walls, backdrop, and poles.  The robot should not move toward the obstacle even if directed to, but may move away from it.  The feature is not tested yet.  Also the projected next position relies on projecting acceleration which has not been implemented.
Field coordinates for starting were added to the driver op so that coordinate based actions could be used, such as collision avoidance.  The coordinates are tested.